Add event loop and fd wrapper for shm transport
The epoll(2) event loop code started out as a copy of the event loop
code for the tcp transport in Gloo. I removed the busy polling on
`epoll_wait(2)` every 10ms and replaced it with an explicit wakeup via
an `eventfd(2)`. The fd wrapper takes care of closing the file
descriptor upon destruction and exposes a few convenience
functions (motivated by the read and write operations on the eventfd
as performed by the event loop).