Initialization of shm connection
The implementation exchanges eventfds over a domain socket between the
peers. After that, it exchanges paths to each other's inbox
ringbuffers. The implementation of transport::Connection::write
directly writes to the peer's inbox and triggers its eventfd. For the
read, the local eventfd is registered with epoll(7). When it is
readable, one or more reads are executed and issue registered
callbacks with pointers to the local inbox memory.