Use std::recursive_mutex in connection
Read or write callbacks are called while holding the connection lock.
This is necessary because they may hold on to some memory in the
shared memory ring buffer. However, these callbacks can in turn call
into the read or write functions, which re-acquires the same mutex.

A better fix here is to use more fine grained locking.