Always run read and write callbacks from event loop thread
Doing so means there is no longer ambiguity which thread ends up
running a callback and whether it is called in place or
asynchronously. It is now always called by the event loop thread and
always asynchronously w.r.t. the call site that registers it.

Separately, ensure that the connection-wide lock is unlocked before we
execute callbacks. This is needed so we can remove the recursive mutex
that we put in place as a stopgap in 0340bf4.