Use shared_ptr for event handlers
Now the event loop can use a weak_ptr to ensure the event handler is
kept alive while the event handler function is executed. This also
means we no longer need to wait for a loop tick upon unregistering an
fd from the event loop. Instead, we can simply reset the weak_ptr and
know that any pending events for that handler are ignored.

Note: registering an fd with the event loop now requires a shared_ptr.
If an object is created through make_shared, the shared_ptr that can
be acquired through shared_from_this isn't valid before all
constructors of an object have finished. Therefore, we take care to
only register an object with the event loop until after it has been
created through make_shared.