File descriptor monitor locks shared_ptr of subject
The lifetime of an file descriptor monitor itself is not enough to
ensure the registered callback can be called. The subject of the user
specified callback must also be kept alive. This commit captures a
weak_ptr to that subject upon construction and only executes the
callback if the weak_ptr can be locked as a shared_ptr.