Use naked pointer to loop on FunctionEventHandler
We capture a shared_ptr to the loop in the lambda in `Loop::monitor`
in order to keep the loop alive from the function event handler
instance. We cannot have the instance store a shared_ptr to the loop
itself, because that would cause a reference cycle when the loop
stores an instance itself.