No longer return future from shm::Loop::run
The returned future would not only signal completion but also hold any
exception that the function could have thrown. If this future was
discarded instead of waited on, an exception could have gone
unnoticed. If in the future we need a run function that returns a
future, make sure it is marked `[[nodiscard]]` to force waiting on the
value of the future.