Use shared_ptr's aliasing constructor instead of a custom deleter
A perhaps little-known feature of shared_ptr is that one can create
a shared_ptr from an arbitrary pointer which shares the control
block of a pre-existing shared_ptr. This is intended for when the
two pointers live in the same memory allocation and thus it should
be kept alive for as long as they both exist.