Use atomic_flag instead of atomic_bool
We didn't need anything from atomic_bool that atomic_flag couldn't
do and I believe the latter should be preferred because it is
simpler (for example, it is guaranteed to always be lock-free).

The atomic_flag contains whether the lock is held by someone. This
commit makes each producer/consumer store whether they are the ones
holding it. This means all the "inTx" safety checks provide more
safety: they don't accidentally pass if someone else is holding the
lock (also, the checks now only use private memory).

This introdues a difference in behavior: we used to have two failure
modes, EBUSY and EAGAIN, with the former one happening when the lock
was already held (by the same process or another one) and the second
one never happening (from my understanding). Now, EAGAIN is returned
if someone else is holding the lock and we need to wait, while EBUSY
is returned if we're holding it and are trying to acquire it
recursively.