Use anonymous shared memory, instead of linking it to /dev/shm
We used to have the sender allocate some anonymous shared memory,
link it to a unique filesystem path, send that name to the
receiver, which would open that file. At the end, the sender had
to unlink the file.

There are several advantages to have the sender communicate the
file descriptors of that shared memory to the receiver (over the
UNIX domain socket that they're already using for that), rather
than linking it:
- No need to devise unique names, no risk of conflicts
- Automatic permission management: only those with the file
  descriptor can access the memory
- Automatic cleanup: the kernel does refcounting for us and
  deallocates the memory when we're done with it, even if the
  programs crash.
- Simpler handshaking, requiring only one message from each side
  (thus one less roundtrip), and a much smaller total size.
- Much less code. By far. Also, no need for the filesystem
  polyfill anymore.