Use "WithCachedSizes" version of Protobuf serialization
Serializing a protobuf message requires knowing the size of each field,
recursively. This is something we're already computing when we ask for
the global size of the entire message, and protobuf caches these values
and offers a serialization method that takes advantage of them rather
than re-computing them. It should thus be a bit faster.