Use "WithCachedSizes" version of Protobuf serialization (#39)
Summary:
Serializing a protobuf message requires knowing the size of each field,
recursively. This is something we're already computing when we ask for
the global size of the entire message, and protobuf caches these values
and offers a serialization method that takes advantage of them rather
than re-computing them. It should thus be a bit faster.
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/39

Differential Revision: D19660500

Pulled By: lerks

fbshipit-source-id: 07739f5f9f6efefbbf468efded8b98176738cc99