Call join in destructors if it wasn't called yet (#37)
Summary:
We typically use shared_ptrs as "loose" unique_ptrs, which have one
"moral" owner, but for which we need to be able to hold additional
references for example in callbacks. (These references are stored as
weak_ptrs to avoid objects being kept artificially alive and thus
preventing leaks, and are only converted back to shared_ptrs when
needed). Because of this, however, when the "moral" owner resets its
shared_ptr it isn't guaranteed to immediately trigger the destruction
of the object, which may be delyed by these other short-lived
references. The join method is thus used to synchronously wait for
everything to be shut down, making the object "dead", even if the
object isn't destroyed yet.

Until now the user was always supposed to call the join method
manually just before destorying its pointer. However in cases where
the scope exited early, for example due to exceptions, this caused
issues. Namely, the exception handling would call the object's
destructor without having called join, which would cause an assert
to fail and thus a runtime_error to be raised from the destructor,
which would then cause the object's members to be destroyed, which
include a condition_variable on which a *non-joined* thread was
waiting. This manifested itself as a data race under tsan. This
hid the original error and made debugging very hard.
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/37

Differential Revision: D19660490

Pulled By: lerks

fbshipit-source-id: 1c742bdc4d2d774350a70a03203c222b7431bc54