Allow to join UV loop with outstanding handles/requests
Summary:
A deadlock occurred when uv::Loop::join was called with outstanding
handles and requests (for example due to delayed destruction) because
joining closed the async handle and thus when the destructors of these
objects wanted to close their handles/requests they would schedule the
close operation using uv::Loop::defer, which would append them to the
queue and would call wakeup, which would send an event on the async
handler, which was however closed and would thus not wake up the loop.
The loop therefore stayed alive forever, with some unclosed handle or
request and with a function scheduled that would never be run.

This is the same problem that #38 fixed for the SHM loop.

The fix here is to unref the async handler, instead of closing it, so
that it stays active and keeps firing and waking up the loop, but it
won't keep the loop alive if it's the last remaining handle. It needs
to be closed afterwards however, with a second uv_run call.

Reviewed By: pietern

Differential Revision: D19690780

fbshipit-source-id: f360c7b9b95f3ead32de8e1d0556e62966928229