Rectify semantics of UV's accept
Summary:
Two things were off:
- once an accept callback fires, it becomes disarmed and must be rearmed
- firings that occur while the callback is disarmed are stashed and
  triggered as soon as it's rearmed, rather than being discarded

Contrary to SHM, we don't have yet the ability to disable the lower-level
UV callback when the user callback is disarmed. So we'll keep getting
notified of new connections even if we don't know what to do with them
and don't want them. Thus we must store them somewhere. This is what
RearmableCallback is for.

PS: it turns out SHM's one was also slightly off, contemplating the
possibility of the callback to be disarmed, when this couldn't be
the case.

Reviewed By: pietern

Differential Revision: D19690869

fbshipit-source-id: 3b9df7b3ea8f4bb341a4ca23a443f382551fd326