Have the client send the brochure instead
Summary:
This avoids us registering a connection request that we then may not need. It also simplifies the logic in the server because it now needs to wait for only one thing at a time rather than two (the brochure answer _and_ the new connection).

This relies on the domain descriptor comparison being symmetric (it now is, because it's an equality, let's hope it will always be that).

The client hello and the brochure are sent back-to-back.

Reviewed By: pietern

Differential Revision: D19725570

fbshipit-source-id: 77b886ecc9d346c5e3dcab151b1e50535eaeba23