Call pending callbacks when pipe is destroyed
Summary: The title of the diff is the primary objective, which is achieved in lines 118-119, which were the first ones I wrote. Problem: you can't call shared_from_this in the destructor, thus you can't keep the messages on the pipe and give the callback a pointer to the pipe so that it can access the messages. Therefore the messages must go in the closure of the callback, but they can't because they're non-copyable. Thus all this other stuff.

Reviewed By: pietern

Differential Revision: D19855480

fbshipit-source-id: ca59ce84c670e8d59035d802c1c540c79a19b893