Store errors by value, rather than a reference to them
Summary: The template parameters used to define the type of the data we'd store, but we were doing the mistake of storing references to errors that may become invalid by the time we accessed them. Now the template parameters define the type of the data the function takes, and we store those types with the references removed. A bit of template magic is needed.

Reviewed By: pietern

Differential Revision: D19855479

fbshipit-source-id: d39024f45aa1dded89d96ae52785af89ed1e0ba5