Make rearmable callback store queue of callbacks
Summary:
This is a (perhaps subtle) design issue. Arming (say) a readDescriptor callback which was already armed ended up destroying the previous callback. This was breaking the consistency of the API:
- you can call read and write and other functions multiple times in a row and they "queue"
- we otherwise _never_ destruct a callback without firing it first

Moreover, allowing to queue these callbacks can optimize a few "backpressure" cases: if we were to (say) register and un-register a listening socket for readability each time the accept callback got armed and un-armed we may do it a lot, by allowing the user to always keep at least at least two callbacks queued we'd allow it to make sure the socket never gets unregistered.

Reviewed By: pietern

Differential Revision: D19936116

fbshipit-source-id: 75b2b60e9fd21a89ba8c9702dbb1e16378726cb5