Don't double-register the descriptor read
Summary: We used to register a read for a descriptor twice: after registering the read for the previous message's payload, and after completing that read. This lead us to expect two descriptors in a row, rather than one descriptor followed by a payload. This only showed up when sending two or more messages, and was thus not caught by our tests, which only sent one. I updated the tests in D19936115 (and made sure that those tests were failing before this diff).

Reviewed By: pietern

Differential Revision: D19936117

fbshipit-source-id: b7c19b792fd0f01c4597cc118474bca91c7bdb2f