Reorganize transport test code
Summary: Define transport tests in a header in the `transports` directory without any mention of Uv, Shm or any concrete implementation, and then in each `uv`, `shm`, ... directory import that header and "call it" with the actual implementation. This way adding one implementation won't require changes to the generic code.

Reviewed By: beauby

Differential Revision: D19939315

fbshipit-source-id: b7bdcc969c1cc61cbff12bbc289f28c22061f2b6