Allow transports to specialize protobuf read
Summary:
The connections have a helper method to read a protobuf, which is implemented in the abstract transport by wrapping the abstract buffer read method. This protobuf read helper was templated. Unfortunately this means that subclasses cannot redefine it (templates and virtual methods don't go along well). There are cases where subclasses may want to specialize it (e.g., shm could read a protobuf directly from the stream, without an intermediate copy in a temporary buffer; we already do this when writing).

One way to make this possible is avoiding templates and instantiate a empty protobuf before calling read, pass a reference to it to read, which reads into it. This way read can just take a reference to the superclass, rather than a template parameter.

Reviewed By: patricklabatut

Differential Revision: D20066332

fbshipit-source-id: 59f2130e5ae7db5a6d21ef74b9e5a8352c7f2b1c