Use raw pointers for message
Summary:
And replace rvalue refs in signatures with a deleted copy constructor / assignment.

We used unique_ptrs to highlight the ownership model: when passing a message to a pipe, we pass ownership of the data to it, until we get the message back. This sounded good on paper but had a couple of issues: when the memory wasn't uniquely owned to start with (say, it came from a pre-existing tensor) it was odd to wrap it in a unique_ptr: one always had to release it at the end to avoid accidentally deallocating it. When forgetting to release, the memory could accidentally get freed but, what's worse, it could be freed incorrectly, namely using delete[] instead of free. This would then trigger another swath of hard-to-debug memory issues. (ASAN would catch that, though). Also, the delete[] may actually fail, in which case the entire program would crash (this was actually a good indicator of a message going out of scope when unintended).

At the end of the day, it may be simpler to just use raw pointer and let the user deal with ownership and deallocation the way they prefer.

I do however still have mixed feelings about this, as all the above issues occasionally turned out to be helpful in debugging another problem in the code.

Reviewed By: jiayisuse

Differential Revision: D20066333

fbshipit-source-id: 45e2106899a2ab980d23630356c896f62dc14c23