Call pipe read and write callbacks inline
Summary:
We used to have a thread in the context that took care of firing all the callbacks of all its pipes. This was done primarily in order to avoid deadlocks (if the pipe was holding a lock while calling a callback, and the callback called in turn a method of the pipe which would try to acquire the same lock). This thread also allowed us to better isolate the user code, so that if it misbehaved it would not affect the functionality of TensorPipe. Moreover, it provided a clean and simple treading model to the user, hiding all the implementation details.

However, this led to a quite poor latency. While we were measuring a RTT latency of 3us with the "raw" SHM transport, this went up to 23us when using the full pipe. After some exploration, it seems the biggest culprit was the use of this external thread. Removing it, and calling callbacks "inline" (i.e., from the event loop threads, the reactor threads, ... wherever) we get the latency down to 6us.

I'm not super happy to remove this thread, as it did provide several advantages, but we cannot sacrifice performance. If any user wants to extra safeguards, they can re-implement this defer-to-thread mechanism in their own code.

Observe that even when calling callbacks inline, there are some guarantees that we can still honor. An important one is making sure that we call the callbacks in the same order as they were schedules, which boils down to preserving the order of messages as they cross the pipe. We do that by explicitly tracking the sequence number of the messages and their callbacks, and only invoking a callbacks when all those with a lower sequence number have finished.

Note that two problems remain:
- Callbacks of different types (reads vs writes) may be called at the same time. This can be fixed by an extra mutex.
- The pipe could still deadlock for the original reason if we go one step further: the pipe calls a callback, which calls a pipe method, which calls another callback *inline*. Unfortunately I don't see any way of fixing it this while keeping up our promises of not calling two callbacks at the same time and of not using external threads. Luckily for us, the condition above happens only when the pipe is in the error state. Also, luckily such a condition can be detected. So we could raise an exception in those cases, or use an external thread only in those cases. Both solutions would be alright I guess. I'll implement one of them in a later diff.

Reviewed By: jiayisuse

Differential Revision: D20117741

fbshipit-source-id: aa68c457f7217f67ccf4f443e6e8e8bcc18c2ac2