Give internal methods access to the lock
Summary:
This is basically just a lot of piping.

For a couple of reasons (reducing latencies and avoiding deadlocks) we need to be able to temporarily give up the pipe's lock when calling into a connection or a channel or when calling a user callback. So far this wasn't possible because the lock was acquired at the entry point and the inner methods didn't have access to it. Now we pass around a reference to it.

It makes the code more unwieldy but my hope is that this is just temporary. We're aligning more and more the pipe, the listener and the channel, and once they're all similar we should be able to extract most of that repetition into a helper class (using shiny features like templates and pointers-to-member-functions) and streamline all of this.

Reviewed By: jiayisuse

Differential Revision: D20096965

fbshipit-source-id: 4daf3cd6adea1559de9efbb730b24d8816b5ba49