Enforce proper order in read descriptor callbacks
Summary: This diff completes the previous one, by using sequence numbers to make sure the read descriptor callbacks are invoked in the order they were scheduled. In an effort to get rid of the RearmableCallback (which would have made this more complicated), this diff also implements backpressure (which was something we wanted to do regardless at some point). What this means is that the pipe will only register a read operation with the connection when it is expecting a descriptor or some data. If the pipe isn't expecting anything (because the user paused the pipe), then the pipe won't ask the underlying connection to read anything. The connection can propagate this "pausing" all the way to the kernel, thus avoiding useless wakeups and performing better.

Reviewed By: jiayisuse

Differential Revision: D20117743

fbshipit-source-id: 3e9bc64f0a1d2e760ccd685abe7b8fed518c2cb9