Use the same lock for the pipe/listener and RearmableCallback
Summary: We need to create two versions of RearmableCallback, one with its own lock and one with an external lock. This leads to some duplication. I may give a try to a single (even more) generic version supporting both, but for now this was simpler. Note that the pipe will stop using the RearmableCallback in a subsequent diff.

Reviewed By: jiayisuse

Differential Revision: D20117742

fbshipit-source-id: 1e68f325011e1e7eb8f2f92ffca1a1e2c894d716