Fix CMake build of transport and channel tests
Summary: The generic tests for transports and channels (the ones that are written against the abstract base class) were not picked up in their concrete instantiations. So I revamped the CMake rules by adding two functions (`add_tensorpipe_function_test` and `add_tensorpipe_channel_test`) which make sure to always correctly link those tests. We already had a function of that type (`add_tensorpipe_test`), which was needed to ensure that the sigpipe ignorer was always linked in. I had troubles stacking two levels of link-whole, so instead I switched to using object libraries: this way the object files are kept around and only linked in at the very end.

Reviewed By: beauby

Differential Revision: D20189074

fbshipit-source-id: 42dc885431f7f939200049955a06dbd1ea4a35d9