Call the listener's callbacks inline
Summary:
Basically, do to the listener what we did to the pipe in D20117741.

In order to avoid a false positive lock-order-inversion warning from TSAN, the listener must construct the pipe while it's not holding the lock. The best place to do so (and avoid race conditions) is when we already release the lock to call the callback. Therefore we now pass around the transport and the connection (the "ingredients" of the pipe) rather than a built pipe, until the very last moment.

Reviewed By: beauby

Differential Revision: D20188767

fbshipit-source-id: 8d014c6c3784a367403cb8bf1be158cf6a28f9d3