Tag SHM methods that are run from reactor
Summary:
The SHM transport is effectively a single-threaded library, where most of the work is performed from within the reactor thread. This is far from obvious at first glance, as it requires tracking down the call graph. To make it more self-evident I appended `FromReactor` to all methods of the connection that are supposed to be run by the reactor thread only. I also added a TP_DCHECK for that, so that we can easily catch any violations. (No violations have been spotted by the test suite).

The only way for some code in another thread (different from the reactor) to run code in the reactor is by passing through loop->deferToReactor. This results in that work being encapsulated in a lambda and enqueued. In the connection, this happens in the methods starting with `trigger`, namely `triggerProcessWriteOperations` and `triggerProcessReadOperations`. In my opinion it should also be supposed to happen in the close method, to avoid races.

Note that the places where we need to perform this deferral to the reactor happen to be exactly the methods of the connection's public interface (i.e., the abstract base class): namely the read method, the write method and the destructor. I think this suggests that the technique is generalizable to every transport that is driven by a serial looping thread.

Reviewed By: jiayisuse

Differential Revision: D20248155

fbshipit-source-id: fabad3bcc67ef7622eb1414edead3e6ddc7d1d48