Avoid runIfAlive when possible
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

The runIfAlive helper was there to help us avoid callbacks keeping objects artificially alive or objects being destroyed in the middle of a callback. Now that the lifetimes of the implementation and its UV handle are tied, we don't need the wrap the callbacks that the former passes to the latter inside runIfAlive.

Reviewed By: beauby

Differential Revision: D20323117

fbshipit-source-id: eacfd61de303d4d5ddc5f0b1d0815f5e2f6ee8c9