Add close callback to UV handles
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

Libuv mandates that memory cleanup of closing handle can only be done once its close callback has been called. Since we now want to decouple the lifetime of the implementation from the connection and couple it with the UV handler, we need to be able to be alerted of that closing happening.

Reviewed By: beauby

Differential Revision: D20312725

fbshipit-source-id: 1e760583ac192cd4bcdb00130fa67da5ccadad81