Remove useless mutexes and locks
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

All the UV handles and the connection/listener implementations run now from within the event loop (we have dchecks to confirm that). They will thus not race against each other and therefore there's no need for mutexes and locks to synchronize them.

Reviewed By: beauby

Differential Revision: D20335011

fbshipit-source-id: ba917917672574b05bfe5b482d4e768e0d43e82a