Cause loop->join to close all connections and listener
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

We decided it's too error prone (actually, deadlock-prone) to have join just wait for everything to terminate without actually trying to shut anything down. So we're changing that.

Reviewed By: beauby

Differential Revision: D20312730

fbshipit-source-id: 98ee9fbae413b1b7f25449c3355881ecaf52f39b