Migrate more UV connection methods to be called from the loop
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

All of the low-level operations on UV handles (readStart, readStop, write) were being deferred to the event loop. This created a weird threading model and a desynchronization between the connection and its handle. This diff moves the boundary between the caller thread and the event loop, by running the connection's methods from within the event loop too.

Reviewed By: beauby

Differential Revision: D20312729

fbshipit-source-id: 8b8858316f76db2d7053eb4838504df8cc168fdb