Close UV connections and listeners asynchronously
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

Here we detach the lifetime of the implementations from the lifetime of the connections/listeners. This means that closing (or even destroying) the connection and listener will schedule closure and destruction of the implementation, but this will be done asynchronously, by the event loop. This is because immediate destruction may race with the event loop if it's accessing some of the connection's data (e.g., if the loop is currently sending some data, and the connection's destructor calls the write callback which frees that memory, we may segfault).

Reviewed By: beauby

Differential Revision: D20312719

fbshipit-source-id: 1bc32767d32625c44e870dca49c595a2695c0d94