Put connection methods that are run by the loop in separate class
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

Since there's now two types of methods on the connection (the private ones, that are supposed to be called from the loop, and the public ones, which just defer to the private ones) let's put the former in a dedicated private nested class. This will help later when we want to detach the lifetime of that class detached from the lifetime of the connection.

Reviewed By: beauby

Differential Revision: D20312723

fbshipit-source-id: 347b61413941d16d737ae6cb266f7f5c37aee7fb