Put listener methods that are run by the loop in separate class
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

This diff does to the listener what the previous diff did to the connection.

Reviewed By: beauby

Differential Revision: D20312727

fbshipit-source-id: 4994453307d67b6d0169e3eb1bbd7e36f39b66b9