Fire all callbacks at close with error
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

Libuv's semantics is that writes are so-called requests, each with its own callback, and all those requests will be cancelled and all those callbacks will be invoked when the connection closes. Reading, on the other hand, is a per-connection callback, and it will not be called with an error if we close the connection. It is thus on us to fire all the pending read callbacks upon closing.

Reviewed By: beauby

Differential Revision: D20335009

fbshipit-source-id: 830b6e3b06680d37b195e749e5449872b98adfc7