Perform UV handle initialization in loop
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

Most of the remaining occurrences of the UV handles deferring operations to the loop is in their initialization (init, bind, connect, ...). This diff changes that, by deferring in the connection/listener and then calling all those functions inline.

Reviewed By: beauby

Differential Revision: D20312720

fbshipit-source-id: 4de5e17684fe06915f8017e8e821f917877a1028