Challenge some assumptions in tests
Summary:
Lots and lots of tests were using the helper to set up a connection, would schedule a write on one side and a read on the other, and then just complete. Since reads and writes are non-blocking, this meant that the helper saw both threads terminate before anything had actually been done. At this point, the helper would join the context.

All of this sort-of worked because the callbacks were capturing the connection, thus keeping it alive. And because joining the context was a "passive" operation, waiting for stuff to finish without actually trying to terminate it.

We'll soon change the latter assumption (join will actively trigger a shutdown), and also the former one we should perhaps make explicit. This diff uses "events" (implemented as promise+future) to explicitly synchronize the operations.

It also adds a couple of helpers that make sure that the callbacks aren't destroyed without being called, and that they are not called inline.

Reviewed By: beauby

Differential Revision: D20335008

fbshipit-source-id: ea73bec7848843d9c2401a18de6bccea7972ee94