Avoid extra allocation for uv_buf_t
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

We were doing a separate heap allocation for the two uv_buf_ts that we need to pass to uv_write. The reasoning was probably that they had to survive until the uv_write callback, even if the WriteOperation, the connection and everything else had gone. This didn't make complete sense (the data we were writing had to remain alive too, which means that the high-level write callback couldn't be called, which means the WriteOperation couldn't be destructed). Since now we're keeping the WriteOperation around anyways, we can store those buffers on

Reviewed By: patricklabatut

Differential Revision: D20335010

fbshipit-source-id: f39c01eae754eff4cdd749e8d3f5284ce2e2dbca