Migrate more UV listener methods to be called from loop
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

This diff does to the listener what the previous diff in the stack did to the connection.

Reviewed By: beauby

Differential Revision: D20312726

fbshipit-source-id: 8523b0fceab94f3c2efb44cf2a7797eea9a4fb9e