Tag UV methods that are run from loop
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

This diff makes it explicit in the code which methods can only be run from the loop (and adds debug checks).

Reviewed By: beauby

Differential Revision: D20312731

fbshipit-source-id: fb9ff31e7cc87301da5286f20a2b95f2396307a2