Remove the last remaining internal deferrals
Summary:
This diff is part of a stack that changes the threading model of the UV transport, so that the public API (read, write, close, ...) only schedules these operations by deferring them to the event loop, which means that all operations and data manipulation is done by the loop, serially. We chose this as a principled approach to avoid race conditions by design.

The very last deferrals done by UV handles were either inside unused methods (noDelay, peerName), which I removed, or inside sockName, for which I moved the deferring to the connection.

Reviewed By: beauby

Differential Revision: D20312728

fbshipit-source-id: 63696206ad3153e0c803697e7d4edbaae5eca34c