Simplify UV listener's accept flow
Summary: We had a complicated machinery in place so that inside the connection callback we could prepare a connection handle and then call accept and pass it a callback that would complete the initialization. However, after the recent redesign of the threading model of the UV transport, the accept is now done inline, and the callback is also called inline. So there's no reason for that machinery.

Reviewed By: beauby

Differential Revision: D20385684

fbshipit-source-id: 1f4437864bb176350c23da40be5bf716143b3ad2