Move around some UV code
Summary:
This diff doesn't change a single line of code.

It just moves code around so that the bulk of connections and listeners is defined in the .cc files, with the .h files just containing the public interface. This has been made possible by the fact that those classes' definitions now only contain a pointer to the implementation object.

The benefits of this is a cleaner header file, and a shallower nesting.

It also moves some member functions from public to private and adds trailing underscores to the private ones.

Reviewed By: beauby

Differential Revision: D20417411

fbshipit-source-id: 84f312d39b8795b24eb752309745f6240f00c456