Make pipe use ephemeral task loops to run operations and callbacks
Summary:
We've recently started refactoring the transports by adopting a structured thread model that should limit race conditions by design: have a single thread (the event loop, the reactor, ...) do all the work, and have all other threads just defer work to this thread, by appending to a queue of tasks.

We want to do the same with the pipe, but there's an obvious problem: the pipe doesn't have its own thread to offload work to. So what this diff does instead is temporarily "borrow" the calling thread (the user thread from which an operation is being scheduled, or the transport thread from which a callback is being fired) to act as the event loop. Such a thread runs the operation/callback it was supposed to run, and then runs all the additional operations/callbacks that were triggered as part of handling the previous one, and so on, until there are no more operations/callbacks to handle. At that point, it just returns, and another thread can come in, handle another operation/callback, and do the same.

Reviewed By: jiayisuse

Differential Revision: D20522369

fbshipit-source-id: 2724df06168ee05fb60a8e1667a168fbeb9cb070