Fix race condition in core tests
Summary:
I saw this as a flake in the ServerPingPongTwice test: the write callback would be called with an error saying that the pipe was closed. This is because we only wait for the reads callbacks to complete before starting to shut down all the pipes. Sometimes the write callback fires after the read one. This diff uses two promised in order to wait for both.

PS: it is possible that this only occurs with the stack I'm currently working on, but the current behavior is still in violation of the API contract and should be fixed.

Reviewed By: beauby

Differential Revision: D20598386

fbshipit-source-id: 4d206845cb999f032dc8366465eb8d6d66a98ff7